+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - added multiselect to task sidebar
 - passing auto approve options and denying execution of bash commands when running Aider via tool
 - added speech to text support using Gemini or OpenAI provider
+- corrected parameter name for internal helper tool when LLM uses invalid tool arguments
 
 ## [0.40.0]